test: update vitest 3 to 4
