+++ packages/editor/test/addPost.shared.ts
@@ -1,4 +1,4 @@
-import { env } from 'cloudflare:test';
+import { env } from 'cloudflare:workers';
 
 import { clearTables, expectJsonResponse, requestJson } from './testUtils';
