+++ pnpm-workspace.yaml
@@ -6,9 +6,9 @@ trustPolicy: no-downgrade
 blockExoticSubdeps: true
 
 catalog:
-  "wrangler": 4.70.0 
-  "@cloudflare/vitest-pool-workers": 0.12.19  
-  "@cloudflare/workers-types": 4.20260305.0
+  "wrangler": 4.73.0 
+  "@cloudflare/vitest-pool-workers": 0.13.0
+  "@cloudflare/workers-types": 4.20260312.1
   "@types/node": 25.5.0
 
 catalogMode: manual
